Fix EE settings menu
