+++ packages/core/admin/admin/src/hooks/useSettingsMenu/utils/adminLinks.js
@@ -1,0 +1,4 @@
+import customAdminLinks from 'ee_else_ce/hooks/useSettingsMenu/utils/customAdminLinks';
+import defaultAdminLinks from './defaultAdminLinks';
+
+export default [...customAdminLinks, ...defaultAdminLinks];
