+++ packages/core/admin/admin/src/hooks/useSettingsMenu/utils/customAdminLinks.js
@@ -1,0 +1,1 @@
+export default [];
