+++ packages/core/admin/ee/admin/hooks/useSettingsMenu/utils/customGlobalLinks.js
@@ -12,17 +12,6 @@ const ssoGlobalRoutes = strapi.features.isEnabled(strapi.features.SSO)
     ]
   : [];
 
-const auditLogsRoutes = [
-  // TODO check if feature enabled
-  {
-    intlLabel: { id: 'global.auditLogs', defaultMessage: 'Audit Logs' },
-    to: '/settings/audit-logs?pageSize=50&page=1&sort=date:DESC',
-    id: 'auditLogs',
-    isDisplayed: false,
-    permissions: adminPermissions.settings.auditLogs.main,
-  },
-];
-
-const customGlobalLinks = [...ssoGlobalRoutes, ...auditLogsRoutes];
+const customGlobalLinks = [...ssoGlobalRoutes];
 
 export default customGlobalLinks;
